# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Vedetta NVR API
  version: "0.1.0"
  description: API for managing cameras, events, recordings, and object detection in the Vedetta Network Video Recorder.
servers:
  - url: /
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /api/cameras/{name}:
    get:
      operationId: getCamera
      summary: Get camera detail
      tags: [cameras]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Camera detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CameraDetail"
        "404":
          description: Camera not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/cameras/{name}/objects:
    post:
      operationId: createObjectFromCameraTrack
      summary: Name a live tracked object on this camera (from the live overlay)
      tags: [objects]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateObjectFromTrackRequest"
      responses:
        "201":
          description: Object created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnownObject"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Camera not found or no live frame available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Object re-identification not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/objects:
    get:
      operationId: listObjects
      summary: List all known objects
      tags: [objects]
      responses:
        "200":
          description: Object list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectListEnvelope"
    post:
      operationId: createObject
      summary: Create a known object from an event detection
      tags: [objects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateObjectRequest"
      responses:
        "201":
          description: Object created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnownObject"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Event snapshot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Object re-identification not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/objects/{id}:
    put:
      operationId: updateObject
      summary: Update a known object's name or match threshold
      tags: [objects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateObjectRequest"
      responses:
        "200":
          description: Object updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusUpdated"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteObject
      summary: Delete a known object
      tags: [objects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Object deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusDeleted"
        "404":
          description: Object not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/objects/{id}/sightings:
    get:
      operationId: listObjectSightings
      summary: List sightings of a known object
      tags: [objects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Sighting list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SightingListEnvelope"
        "400":
          description: Invalid object ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/objects/{id}/crop:
    get:
      operationId: getObjectCrop
      summary: Get cropped image of a known object
      tags: [objects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: ref
          in: query
          description: Optional reference ID to get a specific reference crop
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: JPEG crop image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          description: Crop not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/objects/{id}/references:
    get:
      operationId: listObjectReferences
      summary: List reference images for a known object
      tags: [objects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Reference list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferenceListEnvelope"
        "400":
          description: Invalid object ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: addObjectReference
      summary: Add a reference image to a known object from an event
      tags: [objects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddObjectReferenceRequest"
      responses:
        "201":
          description: Reference added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddObjectReferenceResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Object or event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Object re-identification not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/objects/{id}/thumbnail:
    post:
      operationId: setObjectThumbnail
      summary: Set the avatar/thumbnail of a known object from an event or reference
      tags: [objects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetObjectThumbnailRequest"
      responses:
        "200":
          description: Thumbnail updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetObjectThumbnailResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Object, event, or reference not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/objects/references/{id}:
    delete:
      operationId: deleteObjectReference
      summary: Delete an object reference image
      tags: [objects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Reference deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusDeleted"
  /api/objects/sightings/{id}:
    delete:
      operationId: dismissSighting
      summary: Dismiss an object sighting
      tags: [objects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Sighting dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusDismissed"
        "404":
          description: Sighting not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    StatusDeleted:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [deleted]
    StatusUpdated:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [updated]
...
    StatusDismissed:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [dismissed]
    KnownObject:
      type: object
      required: [id, name, label, created_at]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        label:
          type: string
        crop_path:
          type: string
        match_threshold:
          type: number
          format: double
        created_at:
          type: string
          format: date-time
    CreateObjectRequest:
      type: object
      required: [event_id, name]
      properties:
        event_id:
          type: string
        name:
          type: string
    CreateObjectFromTrackRequest:
      type: object
      required: [track_id, label, name, x1, y1, x2, y2]
      properties:
        track_id:
          type: integer
        label:
          type: string
        name:
          type: string
        x1:
          type: number
          format: float
        y1:
          type: number
          format: float
        x2:
          type: number
          format: float
        y2:
          type: number
          format: float
    UpdateObjectRequest:
      type: object
      properties:
        name:
          type: string
        match_threshold:
          type: number
          format: double
...
    AddObjectReferenceRequest:
      type: object
      required: [event_id]
      properties:
        event_id:
          type: string
...
    CameraDetail:
      type: object
      required: [name, online, has_motion, degraded, ptz, zone_count, recording, sleeping]
      properties:
        name:
          type: string
        online:
          type: boolean
        sleeping:
          type: boolean
          description: >-
            True when an on-demand camera is resting between events. See
            CameraInfo.sleeping.
        has_motion:
          type: boolean
        degraded:
          type: boolean
        degraded_reason:
          type: string
        last_frame:
          type: string
          format: date-time
        ptz:
          type: boolean
        zone_count:
          type: integer
        recording:
          type: boolean
        source_fps:
          type: number
          format: float
          description: Rolling-window source frame rate measured at the detect consumer (frames per second).
        last_connected:
          type: string
          format: date-time
          description: >-
            Time of the camera's last successful RTSP connection. Absent when it
            has not connected since Vedetta started; absent is not the same as
            old, and must not be rendered as an age. For an on-demand camera
            this is the only signal separating one napping between events from
            one that has been unplugged or removed from its bridge, since both
            report the stream as unpublished indefinitely.
        last_seen:
          type: string
          format: date-time
          description: >-
            Time of the last-known frame available for this camera, including
            a cached snapshot restored after restart. Absent when the camera
            has never produced a frame.
        stream_error:
          type: string
          description: >-
            Reason the camera's RTSP source cannot connect, with credentials
            redacted. Present only when the source has failed repeatedly for a
            reason other than its stream not being published, so it is absent
            both while the camera is connected and while an on-demand camera is
            resting between events.
...
    ObjectListEnvelope:
      type: object
      required: [items, total, limit, offset, has_more]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/KnownObject"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
    SightingListEnvelope:
      type: object
      required: [items, total, limit, offset, has_more]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ObjectSighting"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
    ReferenceListEnvelope:
      type: object
      required: [items, total, limit, offset, has_more]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ObjectReference"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
...